requirements.txt
asgiref==3.8.1
beautifulsoup4==4.9.1
bs4==0.0.1
certifi==2020.6.20
chardet==3.0.4
Django==5.1.1
django-admin-thumbnails==0.2.5
django-ckeditor==6.7.1
django-currencies==0.10.1
django-js-asset==2.2.0
django-mptt==0.16.0
django-unfold==0.39.0
idna==2.9
jsonfield==2.1.1
pillow==10.4.0
python-gettext==4.0
pytz==2020.1
requests==2.24.0
setuptools==75.1.0
six==1.15.0
soupsieve==2.0.1
sqlparse==0.3.1
urllib3==1.25.9
